services/auth/src/requirements.txt
alembic==1.15.2
asyncpg==0.30.0
async-fastapi-jwt-auth==0.6.6
fastapi==0.115.11
fastapi-limiter==0.1.6
gunicorn==23.0.0
httpx==0.28.1
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-instrumentation-redis==0.53b1
opentelemetry-instrumentation-sqlalchemy==0.53b1
opentelemetry-sdk==1.32.1
orjson==3.10.15
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic[email]
pydantic-settings==2.8.1
python-jose==3.4.0
redis==5.2.1
SQLAlchemy==2.0.40
typer==0.15.2
uvicorn==0.34.0
werkzeug==3.1.3
